Puppet 4 blockdevices workaround
We don't collect and store the blockdevices fact from nodes because it
changes regularly due to iscsi mounts.  However, init.pp assumes it will
be there, and on puppet4 calling split on undef is a fatal error.  This
will only try to split the blockdevices fact if it already is defined.

This also fixes a typo in the documentation.

Change-Id: I2578f9b0bc64b3da732da7ba9fa7dbe3c6fef2ff